Refactor integer handling to use int64 instead of int32
This commit updates the integer handling throughout the codebase to utilize int64 instead of int32. Changes include modifications to the IntField struct, NewIntField function, serialization logic, and related tests. This enhancement improves the range of integer values supported and ensures consistency across the integer field implementations.

+++ pkg/types/float_test.go
@@ -247,7 +247,7 @@ func TestFloat64Field_Compare_CrossType(t *testing.T) {
 		name     string
 		floatVal float64
 		op       primitives.Predicate
-		intVal   int32
+		intVal   int64
 		expected bool
 	}{
 		{"float equals int", 42.0, primitives.Equals, 42, true},
